[UVS-57] Partial implementation of asynchronous Paho MQTT (embedded-C) client. There are two parts to this: 1) making Paho return immediately after sending network packets, deferring the check for PUBACK, SUBACK, PUBREL, PUBCMP, etc. to the cycle() process, and 2) implementing a Mailbox interface to allow interaction with the MQTT client to be performed indirectly. Neither of these changes affect the existing Paho client behaviour, which is the default. The new behaviour of 1) and 2) can be enabled by building the client with the symbol MQTT_ASYNC defined. The Mailbox interface is accessed using functions MQTTAsyncSubscribe(), MQTTAsyncPublish(), etc., which are only available when this symbol is defined.
All tests currently pass in both configurations, and an additional test suite ("test1a") has
been added for testing the Mailbox interface functions (MQTTAsyncXXX), but I was not able to
actually get these to run since my understanding of the test suite is incomplete.

To build and run the tests, execute travis-build.sh.

Note that no changes have currently been made to the embedded-C++ client. For consistency we can
make the necessary modifications once the changes to the embedded-C client have been solidified.